feat(client): add subagents to create form, share agent form fields
