+++ packages/client/src/components/agent-workspace/agent-form/index.ts
@@ -1,0 +1,11 @@
+export { AgentExecutionLimitsFields } from "./agent-execution-limits-fields";
+export { AgentFormSection } from "./agent-form-section";
+export { AgentModelField } from "./agent-model-field";
+export { AgentSubagentsField } from "./agent-subagents-field";
+export { buildEditWorkspacePixel } from "./pixel";
+export {
+	AGENT_FORM_DEFAULT_VALUES,
+	type AgentFormValues,
+	type ModelEngine,
+	type SubagentEntry,
+} from "./types";
